+++ lib/volt/page/bindings/component_binding.rb
@@ -1,7 +1,7 @@
-require 'volt/page/bindings/template_binding'
+require 'volt/page/bindings/view_binding'
 
 # Component bindings are the same as template bindings, but handle components.
 module Volt
-  class ComponentBinding < TemplateBinding
+  class ComponentBinding < ViewBinding
   end
 end