+++ spec/server/html_parser/view_parser_spec.rb
@@ -164,7 +164,7 @@ describe Volt::ViewParser do
 
   it 'should parse a template' do
     html = <<-END
-    {{ template "/home/temp/path" }}
+    {{ view "/home/temp/path" }}
     END
 
     view = Volt::ViewParser.new(html, 'main/main/main')
@@ -173,7 +173,7 @@ describe Volt::ViewParser do
                                    'html' => "    <!-- $0 --><!-- $/0 -->\n",
                                    'bindings' => {
                                      0 => [
-                                       "lambda { |__p, __t, __c, __id| Volt::TemplateBinding.new(__p, __t, __c, __id, \"main/main/main/body\", Proc.new { [\"/home/temp/path\"] }) }"
+                                       "lambda { |__p, __t, __c, __id| Volt::ViewBinding.new(__p, __t, __c, __id, \"main/main/main/body\", Proc.new { [\"/home/temp/path\"] }) }"
                                      ]
                                    }
                                  })